renaming modules to addons
