+++ addons/cloudevents/cloudevents-quarkus-addon/pom.xml
@@ -21,11 +21,12 @@
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.kie.kogito</groupId>
-    <artifactId>kogito-shared</artifactId>
+    <artifactId>cloudevents</artifactId>
     <version>1.0.0-SNAPSHOT</version>
   </parent>
-  <artifactId>kogito-shared-events-quarkus</artifactId>
-  <name>Kogito :: Shared :: Events :: Quarkus</name>
+  <artifactId>kogito-cloudevents-quarkus-addon</artifactId>
+  <name>Kogito :: Add-Ons :: CloudEvents :: Quarkus</name>
+  <description>Kogito Events handling for CloudEvents AddOn (Quarkus)</description>
 
   <dependencyManagement>
     <dependencies>
