+++ addons/cloudevents/cloudevents-spring-boot-addon/src/main/resources/META-INF/kogito.addon
@@ -1,0 +1,1 @@
+cloudevents
