+++ addons/cloudevents/pom.xml
@@ -21,31 +21,18 @@
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.kie.kogito</groupId>
-    <artifactId>kogito-build-parent</artifactId>
+    <artifactId>addons</artifactId>
     <version>1.0.0-SNAPSHOT</version>
-    <relativePath>../kogito-build-parent/pom.xml</relativePath>
   </parent>
-  <artifactId>kogito-shared</artifactId>
+  <artifactId>cloudevents</artifactId>
   <packaging>pom</packaging>
-  <name>Kogito :: Shared Runtime Components</name>
-  <description>Shared Runtime Components</description>
+  <name>Kogito :: Add-Ons :: Cloud Events</name>
+
+  <description>Kogito Cloud Events</description>
 
   <modules>
-    <module>events-quarkus</module>
-    <module>events-spring</module>
+    <module>cloudevents-quarkus-addon</module>
+    <module>cloudevents-spring-boot-addon</module>
   </modules>
-
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <groupId>org.kie.kogito</groupId>
-        <artifactId>kogito-bom</artifactId>
-        <version>${project.version}</version>
-        <type>pom</type>
-        <scope>import</scope>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
-
 
 </project>
