--- kogito-shared/events-spring/src/main/java/org/kie/kogito/shared/events/spring/SpringKafkaCloudEventEmitter.java
@@ -1,44 +1,0 @@
-/*
- *  Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package org.kie.kogito.shared.events.spring;
-
-import java.util.concurrent.CompletionStage;
-
-import org.kie.kogito.services.event.CloudEventEmitter;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-
-/**
- * Spring implementation delegating to kafka template
- * TODO proper error handling https://issues.redhat.com/browse/KOGITO-3408
- */
-@Component
-public class SpringKafkaCloudEventEmitter implements CloudEventEmitter {
-    @Autowired
-    org.springframework.kafka.core.KafkaTemplate<String, String> emitter;
-
-    public CompletionStage<Void> emit(String e) {
-        return emitter.send("kogito_outgoing_stream", e)
-                .completable()
-                .thenRun(() -> {}); // discard return to comply with the signature
-    }
-
-    public void emit(String topic, String message) {
-        emitter.send(topic, message);
-    }
-}
